Moved ifchanged tests into syntax_tests/test_if_changed.py.
